refactor(frontend): Relocate ClientFactory to configs domain
