--- frontend/app/.server/factories/index.ts
@@ -1,1 +1,0 @@
-export * from './config.factory';
